requirements.txt
pyautogen==0.2.27
docker==7.0.0
streamlit==1.35.0
pandas==2.2.2
matplotlib==3.8.4
python-dotenv==1.0.1